chore: support for tweepy>4.0.0
tweepy 4.0.0 introduced a number of breaking changes. This commit adds
support for those changes, while retaining backwards compatibility with
earlier versions of tweepy.

See https://github.com/tweepy/tweepy/releases/tag/v4.0.0 for the full
set of changes. The ones that affect tidytwitter are

 - `wait_on_rate_limit_notify` is no longer an option
 - s/favorites/get_favorites
 - s/me/verify_credentials
 - the `created_at` attribute of a tweet is no longer timezone naive

The last of these requires a couple of changes in order to retain
backwards compatibility with tweepy<4.0.0. First we ensure the "now"
time is tz aware (see
https://blog.ganssle.io/articles/2019/11/utcnow.html) so >4.0.0 works,
and then we ensure the created_at time is also tz aware so that <4.0.0
still works.